+++ pyteamcity/future/core/utils.py
@@ -1,0 +1,5 @@
+import dateutil.parser
+
+
+def parse_date_string(date_string):
+    return dateutil.parser.parse(date_string)
