+++ pyteamcity/future/core/web_browsable.py
@@ -1,0 +1,6 @@
+import webbrowser
+
+
+class WebBrowsable(object):
+    def open_web_browser(self):
+        webbrowser.open(self.web_url)
